refactor: extract video diffusion worker to separate module
Move init_video_diffusion_worker from main.py to workers/ package:
- Create workers/ directory for worker initialization modules
- Add workers/video_diffusion_worker.py with extracted function
- Update main.py to import from new location
- Add TODO for cancellation support (requires visual_gen hooks)

Addresses review comment about keeping main.py lean.
